Re #0. Removed include of NeXusFile from V3D.h and MatrixWorkspace.h.
