+++ Framework/API/src/MatrixWorkspace.cpp
@@ -29,7 +29,6 @@
 
 using Mantid::Kernel::DateAndTime;
 using Mantid::Kernel::TimeSeriesProperty;
-using NeXus::NXcompression;
 using Mantid::Kernel::Strings::toString;
 
 namespace Mantid {
@@ -1569,121 +1568,6 @@ signal_t MatrixWorkspace::getSignalWithMaskAtCoord(
     const coord_t *coords,
     const Mantid::API::MDNormalization &normalization) const {
   return getSignalAtCoord(coords, normalization);
-}
-
-/** Save the spectra detector map to an open NeXus file.
-* @param file :: open NeXus file
-* @param spec :: list of the Workspace Indices to save.
-* @param compression :: NXcompression int to indicate how to compress
-*/
-void MatrixWorkspace::saveSpectraMapNexus(
-    ::NeXus::File *file, const std::vector<int> &spec,
-    const ::NeXus::NXcompression compression) const {
-  // Count the total number of detectors
-  std::size_t nDetectors = 0;
-  for (auto index : spec) {
-    nDetectors +=
-        this->getSpectrum(static_cast<size_t>(index)).getDetectorIDs().size();
-  }
-
-  if (nDetectors < 1) {
-    // No data in spectraMap to write
-    g_log.warning("No spectramap data to write");
-    return;
-  }
-
-  // Start the detector group
-  file->makeGroup("detector", "NXdetector", 1);
-  file->putAttr("version", 1);
-
-  int numberSpec = int(spec.size());
-  // allocate space for the Nexus Muon format of spctra-detector mapping
-  std::vector<int32_t> detector_index(
-      numberSpec + 1, 0); // allow for writing one more than required
-  std::vector<int32_t> detector_count(numberSpec, 0);
-  std::vector<int32_t> detector_list(nDetectors, 0);
-  std::vector<int32_t> spectra(numberSpec, 0);
-  std::vector<double> detPos(nDetectors * 3);
-  detector_index[0] = 0;
-  int id = 0;
-
-  int ndet = 0;
-  // get data from map into Nexus Muon format
-  for (int i = 0; i < numberSpec; i++) {
-    // Workspace index
-    int si = spec[i];
-    // Spectrum there
-    const auto &spectrum = this->getSpectrum(si);
-    spectra[i] = int32_t(spectrum.getSpectrumNo());
-
-    // The detectors in this spectrum
-    const auto &detectorgroup = spectrum.getDetectorIDs();
-    const int ndet1 = static_cast<int>(detectorgroup.size());
-
-    detector_index[i + 1] = int32_t(
-        detector_index[i] +
-        ndet1); // points to start of detector list for the next spectrum
-    detector_count[i] = int32_t(ndet1);
-    ndet += ndet1;
-
-    std::set<detid_t>::const_iterator it;
-    for (it = detectorgroup.begin(); it != detectorgroup.end(); ++it) {
-      detector_list[id++] = int32_t(*it);
-    }
-  }
-  // Cut the extra entry at the end of detector_index
-  detector_index.resize(numberSpec);
-
-  // write data as Nexus sections detector{index,count,list}
-  std::vector<int> dims(1, numberSpec);
-  file->writeCompData("detector_index", detector_index, dims, compression,
-                      dims);
-  file->writeCompData("detector_count", detector_count, dims, compression,
-                      dims);
-  dims[0] = ndet;
-  file->writeCompData("detector_list", detector_list, dims, compression, dims);
-  dims[0] = numberSpec;
-  file->writeCompData("spectra", spectra, dims, compression, dims);
-
-  // Get all the positions
-  try {
-    Geometry::Instrument_const_sptr inst = this->getInstrument();
-    Geometry::IComponent_const_sptr sample = inst->getSample();
-    if (sample) {
-      Kernel::V3D sample_pos = sample->getPos();
-      for (int i = 0; i < ndet; i++) {
-        double R, Theta, Phi;
-        try {
-          Geometry::IDetector_const_sptr det =
-              inst->getDetector(detector_list[i]);
-          Kernel::V3D pos = det->getPos() - sample_pos;
-          pos.getSpherical(R, Theta, Phi);
-          R = det->getDistance(*sample);
-          Theta = this->detectorTwoTheta(*det) * rad2deg;
-        } catch (...) {
-          R = 0.;
-          Theta = 0.;
-          Phi = 0.;
-        }
-        // Need to get R & Theta through these methods to be correct for grouped
-        // detectors
-        detPos[3 * i] = R;
-        detPos[3 * i + 1] = Theta;
-        detPos[3 * i + 2] = Phi;
-      }
-    } else
-      for (int i = 0; i < 3 * ndet; i++)
-        detPos[i] = 0.;
-
-    dims[0] = ndet;
-    dims.push_back(3);
-    dims[1] = 3;
-    file->writeCompData("detector_positions", detPos, dims, compression, dims);
-  } catch (...) {
-    g_log.error("Unknown error caught when saving detector positions.");
-  }
-
-  file->closeGroup();
 }
 
 /*
