+++ Framework/DataHandling/src/LoadIDFFromNexus.cpp
@@ -12,6 +12,7 @@
 #include <Poco/DOM/NodeIterator.h>
 #include <Poco/File.h>
 #include <Poco/Path.h>
+#include <nexus/NeXusFile.hpp>
 
 using Poco::XML::DOMParser;
 using Poco::XML::Document;
