+++ Framework/DataHandling/src/LoadNXcanSAS.cpp
@@ -20,6 +20,7 @@
 #include <H5Cpp.h>
 #include <Poco/DirectoryIterator.h>
 #include <Poco/Path.h>
+#include <nexus/NeXusFile.hpp>
 #include <type_traits>
 
 using namespace Mantid::Kernel;
