+++ Framework/DataHandling/test/SaveNexusProcessedTest.h
@@ -24,6 +24,10 @@
 #include "MantidKernel/Unit.h"
 #include "MantidKernel/UnitFactory.h"
 #include "MantidDataHandling/LoadRaw3.h"
+#include "MantidGeometry/Instrument.h"
+#include "MantidTestHelpers/FakeObjects.h"
+#include "MantidTestHelpers/ComponentCreationHelper.h"
+#include "MantidTestHelpers/NexusTestHelper.h"
 #include "MantidTestHelpers/WorkspaceCreationHelper.h"
 #include <Poco/File.h>
 #include <Poco/Path.h>
@@ -38,6 +42,7 @@ using namespace Mantid::API;
 using namespace Mantid::Kernel;
 using namespace Mantid::DataHandling;
 using namespace Mantid::DataObjects;
+using namespace Mantid::Geometry;
 using namespace Mantid::NeXus;
 using Mantid::HistogramData::HistogramDx;
 
@@ -84,7 +89,7 @@ public:
     outputFile = doExec(outputFile, useXErrors);
 
     // Assert XError correctness
-    NeXus::File savedNexus(outputFile);
+    ::NeXus::File savedNexus(outputFile);
     savedNexus.openGroup("mantid_workspace_1", "NXentry");
     savedNexus.openGroup("workspace", "NXdata");
 
@@ -462,7 +467,7 @@ public:
     outputFileName = alg.getPropertyValue("Filename");
 
     try {
-      NeXus::File savedNexus(outputFileName);
+      ::NeXus::File savedNexus(outputFileName);
 
       savedNexus.openGroup("mantid_workspace_1", "NXentry");
       savedNexus.openGroup("table_workspace", "NXdata");
@@ -471,7 +476,7 @@ public:
 
       savedNexus.openData("column_1");
 
-      NeXus::Info columnInfo1 = savedNexus.getInfo();
+      ::NeXus::Info columnInfo1 = savedNexus.getInfo();
       TS_ASSERT_EQUALS(columnInfo1.dims.size(), 2);
       TS_ASSERT_EQUALS(columnInfo1.dims[0], 3);
       TS_ASSERT_EQUALS(columnInfo1.dims[1], 4);
@@ -487,7 +492,7 @@ public:
       TS_ASSERT_EQUALS(data1[8], 4);
       TS_ASSERT_EQUALS(data1[11], 7);
 
-      std::vector<NeXus::AttrInfo> attrInfos1 = savedNexus.getAttrInfos();
+      std::vector<::NeXus::AttrInfo> attrInfos1 = savedNexus.getAttrInfos();
       TS_ASSERT_EQUALS(attrInfos1.size(), 6);
 
       if (attrInfos1.size() == 6) {
@@ -509,7 +514,7 @@ public:
 
       savedNexus.openData("column_2");
 
-      NeXus::Info columnInfo2 = savedNexus.getInfo();
+      ::NeXus::Info columnInfo2 = savedNexus.getInfo();
       TS_ASSERT_EQUALS(columnInfo2.dims.size(), 2);
       TS_ASSERT_EQUALS(columnInfo2.dims[0], 3);
       TS_ASSERT_EQUALS(columnInfo2.dims[1], 2);
@@ -523,7 +528,7 @@ public:
       TS_ASSERT_EQUALS(data2[3], 2.5);
       TS_ASSERT_EQUALS(data2[5], 0.0);
 
-      std::vector<NeXus::AttrInfo> attrInfos2 = savedNexus.getAttrInfos();
+      std::vector<::NeXus::AttrInfo> attrInfos2 = savedNexus.getAttrInfos();
       TS_ASSERT_EQUALS(attrInfos2.size(), 6);
 
       if (attrInfos2.size() == 6) {
@@ -633,7 +638,7 @@ public:
     // Get full output file path
     outputFileName = alg.getPropertyValue("Filename");
 
-    NeXus::File savedNexus(outputFileName);
+    ::NeXus::File savedNexus(outputFileName);
 
     savedNexus.openGroup("mantid_workspace_1", "NXentry");
     savedNexus.openGroup("table_workspace", "NXdata");
@@ -697,13 +702,13 @@ public:
     {
       savedNexus.openData("column_9");
 
-      NeXus::Info columnInfo = savedNexus.getInfo();
+      ::NeXus::Info columnInfo = savedNexus.getInfo();
       TS_ASSERT_EQUALS(columnInfo.dims.size(), 2);
       TS_ASSERT_EQUALS(columnInfo.dims[0], 3);
       TS_ASSERT_EQUALS(columnInfo.dims[1], 9);
       TS_ASSERT_EQUALS(columnInfo.type, NX_CHAR);
 
-      std::vector<NeXus::AttrInfo> attrInfos = savedNexus.getAttrInfos();
+      std::vector<::NeXus::AttrInfo> attrInfos = savedNexus.getAttrInfos();
       TS_ASSERT_EQUALS(attrInfos.size(), 3);
 
       if (attrInfos.size() == 3) {
@@ -767,7 +772,7 @@ public:
     // Get full output file path
     outputFileName = alg.getPropertyValue("Filename");
 
-    NeXus::File savedNexus(outputFileName);
+    ::NeXus::File savedNexus(outputFileName);
 
     savedNexus.openGroup("mantid_workspace_1", "NXentry");
     savedNexus.openGroup("table_workspace", "NXdata");
@@ -782,13 +787,13 @@ public:
     {
       savedNexus.openData("column_2");
 
-      NeXus::Info columnInfo = savedNexus.getInfo();
+      ::NeXus::Info columnInfo = savedNexus.getInfo();
       TS_ASSERT_EQUALS(columnInfo.dims.size(), 2);
       TS_ASSERT_EQUALS(columnInfo.dims[0], 3);
       TS_ASSERT_EQUALS(columnInfo.dims[1], 1);
       TS_ASSERT_EQUALS(columnInfo.type, NX_CHAR);
 
-      std::vector<NeXus::AttrInfo> attrInfos = savedNexus.getAttrInfos();
+      std::vector<::NeXus::AttrInfo> attrInfos = savedNexus.getAttrInfos();
       TS_ASSERT_EQUALS(attrInfos.size(), 3);
 
       if (attrInfos.size() == 3) {
@@ -856,16 +861,31 @@ public:
     AnalysisDataService::Instance().remove("testSpace");
   }
 
+  void test_nexus_spectraMap() {
+    NexusTestHelper th(true);
+    th.createFile("MatrixWorkspaceTest.nxs");
+    auto ws = makeWorkspaceWithDetectors(100, 50);
+    std::vector<int> spec;
+    for (int i = 0; i < 100; i++) {
+      // Give some funny numbers, so it is not the default
+      ws->getSpectrum(size_t(i)).setSpectrumNo(i * 11);
+      ws->getSpectrum(size_t(i)).setDetectorID(99 - i);
+      spec.push_back(i);
+    }
+    SaveNexusProcessed alg;
+    TS_ASSERT_THROWS_NOTHING(alg.saveSpectraMapNexus(*ws, th.file, spec););
+  }
+
 private:
-  void doTestColumnInfo(NeXus::File &file, int type,
+  void doTestColumnInfo(::NeXus::File &file, int type,
                         const std::string &interpret_as,
                         const std::string &name) {
-    NeXus::Info columnInfo = file.getInfo();
+    ::NeXus::Info columnInfo = file.getInfo();
     TSM_ASSERT_EQUALS(name, columnInfo.dims.size(), 1);
     TSM_ASSERT_EQUALS(name, columnInfo.dims[0], 3);
     TSM_ASSERT_EQUALS(name, columnInfo.type, type);
 
-    std::vector<NeXus::AttrInfo> attrInfos = file.getAttrInfos();
+    std::vector<::NeXus::AttrInfo> attrInfos = file.getAttrInfos();
     TSM_ASSERT_EQUALS(name, attrInfos.size(), 3);
 
     if (attrInfos.size() == 3) {
@@ -880,16 +900,16 @@ private:
     }
   }
 
-  void doTestColumnInfo2(NeXus::File &file, int type,
+  void doTestColumnInfo2(::NeXus::File &file, int type,
                          const std::string &interpret_as,
                          const std::string &name, int dim1) {
-    NeXus::Info columnInfo = file.getInfo();
+    ::NeXus::Info columnInfo = file.getInfo();
     TSM_ASSERT_EQUALS(name, columnInfo.dims.size(), 2);
     TSM_ASSERT_EQUALS(name, columnInfo.dims[0], 3);
     TSM_ASSERT_EQUALS(name, columnInfo.dims[1], dim1);
     TSM_ASSERT_EQUALS(name, columnInfo.type, type);
 
-    std::vector<NeXus::AttrInfo> attrInfos = file.getAttrInfos();
+    std::vector<::NeXus::AttrInfo> attrInfos = file.getAttrInfos();
     TSM_ASSERT_EQUALS(name, attrInfos.size(), 6);
 
     if (attrInfos.size() == 6) {
@@ -905,7 +925,7 @@ private:
   }
 
   template <typename T>
-  void doTestColumnData(const std::string &name, NeXus::File &file,
+  void doTestColumnData(const std::string &name, ::NeXus::File &file,
                         const T expectedData[], size_t len = 3) {
     std::vector<T> data;
     file.getData(data);
@@ -975,6 +995,34 @@ private:
     TS_ASSERT_THROWS_NOTHING(algToBeTested.execute());
     TS_ASSERT(algToBeTested.isExecuted());
     return outputFile;
+  }
+
+  /** Create a workspace with numSpectra, with
+   * each spectrum having one detector, at id = workspace index.
+   * @param numSpectra
+   * @return
+   */
+  boost::shared_ptr<MatrixWorkspace>
+  makeWorkspaceWithDetectors(size_t numSpectra, size_t numBins) const {
+    boost::shared_ptr<MatrixWorkspace> ws2 =
+        boost::make_shared<WorkspaceTester>();
+    ws2->initialize(numSpectra, numBins, numBins);
+
+    auto inst = boost::make_shared<Instrument>("TestInstrument");
+    // We get a 1:1 map by default so the detector ID should match the spectrum
+    // number
+    for (size_t i = 0; i < ws2->getNumberHistograms(); ++i) {
+      // Create a detector for each spectra
+      Detector *det =
+          new Detector("pixel", static_cast<detid_t>(i), inst.get());
+      det->setShape(
+          ComponentCreationHelper::createSphere(0.01, V3D(0, 0, 0), "1"));
+      inst->add(det);
+      inst->markAsDetector(det);
+      ws2->getSpectrum(i).addDetectorID(static_cast<detid_t>(i));
+    }
+    ws2->setInstrument(inst);
+    return ws2;
   }
 
   std::string outputFile;
