+++ Framework/Geometry/inc/MantidGeometry/Crystal/MatrixVectorPairParser.h
@@ -7,6 +7,7 @@
 
 #include "MantidKernel/Exception.h"
 #include <functional>
+#include <map>
 #include <boost/spirit/include/qi.hpp>
 
 namespace Mantid {
