+++ Framework/Geometry/inc/MantidGeometry/Crystal/PointGroup.h
@@ -10,6 +10,7 @@
 #include <vector>
 #include <string>
 #include <set>
+#include <map>
 
 #include "MantidGeometry/Crystal/SymmetryOperation.h"
 #include "MantidGeometry/Crystal/Group.h"
