+++ Framework/Geometry/inc/MantidGeometry/Surfaces/Line.h
@@ -4,6 +4,7 @@
 #include "MantidGeometry/DllConfig.h"
 #include "MantidKernel/V3D.h"
 #include <complex>
+#include <list>
 
 namespace Mantid {
