+++ Framework/Kernel/src/V3D.cpp
@@ -9,6 +9,7 @@
 #include "MantidKernel/Exception.h"
 #include "MantidKernel/Quat.h"
 #include <boost/math/common_factor.hpp>
+#include <nexus/NeXusFile.hpp>
 
 namespace Mantid {
 namespace Kernel {
